everyplace/pyproject.toml
[tool.poetry]
name = "myplace"
version = "0.1.0"
description = ""
authors = ["young <cocoujy@naver.com>"]

[tool.poetry.dependencies]
python = "^3.11.4"
django = "^4.2.4"
djangorestframework = "^3.14.0"
pillow = "^10.0.0"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.2.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.7"
dj-rest-auth = "2.1.4"
django-allauth = "0.51.0"
boto3 = "^1.28.35"
django-storages = "^1.13.2"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
channels = "3.0.5"
drf-spectacular = "^0.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
